+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for interval-patterns
 
+## 0.5.1.0 - 2022-08-04
+
+* improve performance of `Layers.fromList`
+
 ## 0.5.0.0 - 2022-08-04
 
 * fix direction of `diffUTCTime` in `erlangs`